requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cors==1.0.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
future==1.0.0
gevent==25.5.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.78.1
pydantic==2.11.4
pydantic_core==2.33.2
pyodbc==5.2.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
setuptools==80.7.1
sniffio==1.3.1
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
